add syscall sleep and time

+++ kernel/global.c
@@ -4,7 +4,6 @@
 //          Zixuan Liu <zixuanl@andrew.cmu.edu>
 //	    Jianan Lu <jiananl@andrew.cmu.edu>
 // Date:    10/23/2013
-
 #include <types.h>
 
 uint32_t oldsp = 0;
@@ -14,4 +13,5 @@ uint32_t link = 0 ;
 int32_t value = 0 ;
 int32_t ExitStatus = 0;
 
-uint64_t system_time = 0;
+uint32_t system_time = 0;
+
